+++ backend/modules/faq/engine/model.php
@@ -46,7 +46,7 @@ class BackendFaqModel
 		);
 
 		// delete extra
-		$db->delete('pages_extras', 'id = ? AND module = ? AND type = ? AND action = ?', array($extra['id'], $extra['module'], $extra['type'], $extra['action']));
+		$db->delete('modules_extras', 'id = ? AND module = ? AND type = ? AND action = ?', array($extra['id'], $extra['module'], $extra['type'], $extra['action']));
 
 		// delete the record
 		$db->delete('faq_categories', 'id = ?', array((int) $id));
@@ -206,17 +206,17 @@ class BackendFaqModel
 			'hidden' => 'N',
 			'sequence' => $db->getVar(
 				'SELECT MAX(i.sequence) + 1
-				 FROM pages_extras AS i
+				 FROM modules_extras AS i
 				 WHERE i.module = ?', array('faq')
 			)
 		);
 		if(is_null($extra['sequence'])) $extra['sequence'] = $db->getVar(
 			'SELECT CEILING(MAX(i.sequence) / 1000) * 1000
-			 FROM pages_extras AS i'
+			 FROM modules_extras AS i'
 		);
 
 		// insert extra
-		$item['extra_id'] = $db->insert('pages_extras', $extra);
+		$item['extra_id'] = $db->insert('modules_extras', $extra);
 		$extra['id'] = $item['extra_id'];
 
 		// insert and return the new id
@@ -229,7 +229,7 @@ class BackendFaqModel
 			'language' => $item['language'],
 			'edit_url' => BackendModel::createURLForAction('edit') . '&id=' . $item['id'])
 		);
-		$db->update('pages_extras', $extra, 'id = ? AND module = ? AND type = ? AND action = ?', array($extra['id'], $extra['module'], $extra['type'], $extra['action']));
+		$db->update('modules_extras', $extra, 'id = ? AND module = ? AND type = ? AND action = ?', array($extra['id'], $extra['module'], $extra['type'], $extra['action']));
 
 		// return the new id
 		return $item['id'];
@@ -289,7 +289,7 @@ class BackendFaqModel
 		);
 
 		// update extra
-		$db->update('pages_extras', $extra, 'id = ? AND module = ? AND type = ? AND action = ?', array($extra['id'], $extra['module'], $extra['type'], $extra['action']));
+		$db->update('modules_extras', $extra, 'id = ? AND module = ? AND type = ? AND action = ?', array($extra['id'], $extra['module'], $extra['type'], $extra['action']));
 
 		// update category
 		return $db->update('faq_categories', $item, 'id = ? AND language = ?', array($item['id'], $item['language']));