+++ backend/modules/groups/actions/edit.php
@@ -130,8 +130,8 @@ class BackendGroupsEdit extends BackendBaseActionEdit
 		// create filter with modules which may not be displayed
 		$filter = array('authentication', 'error', 'core');
 
-		// get all active filtered modules
-		$modules = array_diff(BackendModel::getModules(true), $filter);
+		// get all modules
+		$modules = array_diff(BackendModel::getModules(), $filter);
 
 		// loop all modules
 		foreach($modules as $module)
@@ -231,7 +231,8 @@ class BackendGroupsEdit extends BackendBaseActionEdit
 	 */
 	private function getWidgets()
 	{
-		$modules = BackendModel::getModules(true);
+		// get all modules
+		$modules = BackendModel::getModules();
 
 		// loop all modules
 		foreach($modules as $module)