+++ backend/modules/locale/ajax/save_translation.php
@@ -29,7 +29,7 @@ class BackendLocaleAjaxSaveTranslation extends BackendBaseAJAXAction
 
 		// get parameters
 		$language = SpoonFilter::getPostValue('language', array_keys($possibleLanguages), null, 'string');
-		$module = SpoonFilter::getPostValue('module', BackendModel::getModules(false), null, 'string');
+		$module = SpoonFilter::getPostValue('module', BackendModel::getModules(), null, 'string');
 		$name = SpoonFilter::getPostValue('name', null, null, 'string');
 		$type = SpoonFilter::getPostValue('type', BackendModel::getDB()->getEnumValues('locale', 'type'), null, 'string');
 		$application = SpoonFilter::getPostValue('application', array('backend', 'frontend'), null, 'string');