+++ backend/modules/locale/engine/model.php
@@ -362,7 +362,7 @@ class BackendLocaleModel
 	public static function getNonExistingBackendLocale($language)
 	{
 		$tree = self::getTree(BACKEND_PATH);
-		$modules = BackendModel::getModules(false);
+		$modules = BackendModel::getModules();
 
 		// search fo the error module
 		$key = array_search('error', $modules);
@@ -377,7 +377,7 @@ class BackendLocaleModel
 		foreach((array) $lbl as $label) $used['lbl'][$label] = array('files' => array('<small>used in navigation</small>'), 'module_specific' => array());
 
 		// get labels from table
-		$lbl = (array) BackendModel::getDB()->getColumn('SELECT label FROM pages_extras');
+		$lbl = (array) BackendModel::getDB()->getColumn('SELECT label FROM modules_extras');
 		foreach((array) $lbl as $label) $used['lbl'][$label] = array('files' => array('<small>used in database</small>'), 'module_specific' => array());
 
 		// loop files
@@ -1000,8 +1000,8 @@ class BackendLocaleModel
 		// paths that should be ignored
 		$ignore = array(BACKEND_CACHE_PATH, BACKEND_CORE_PATH . '/js/tiny_mce', FRONTEND_CACHE_PATH);
 
-		// get active modules
-		$activeModules = BackendModel::getModules(true);
+		// get modules
+		$modules = BackendModel::getModules();
 
 		// get the folder listing
 		$items = SpoonDirectory::getList($path, true, array('.svn', '.git'));
@@ -1017,7 +1017,7 @@ class BackendLocaleModel
 			else $moduleName = substr($path, $start, ($end - $start));
 
 			// don't go any deeper
-			if(!in_array($moduleName, $activeModules)) return $tree;
+			if(!in_array($moduleName, $modules)) return $tree;
 		}
 
 		foreach($items as $item)
@@ -1137,7 +1137,7 @@ class BackendLocaleModel
 
 		// possible values
 		$possibleApplications = array('frontend', 'backend');
-		$possibleModules = BackendModel::getModules(false);
+		$possibleModules = BackendModel::getModules();
 		$possibleLanguages = array('frontend' => array_keys(BL::getWorkingLanguages()), 'backend' => array_keys(BL::getInterfaceLanguages()));
 		$possibleTypes = array();