+++ install/engine/step.php
@@ -30,7 +30,7 @@ class InstallerStep
 	 * @var array
 	 */
 	protected $modules = array(
-		'required' => array('locale', 'settings', 'users', 'groups', 'pages', 'search', 'content_blocks', 'tags'),
+		'required' => array('locale', 'settings', 'users', 'groups', 'extensions', 'pages', 'search', 'content_blocks', 'tags'),
 		'hidden' => array('authentication', 'dashboard', 'error'),
 		'optional' => array()
 	);